Moved funnel series into a subfolder.
