+++ js/Series/Funnel/FunnelSeries.js
@@ -23,13 +23,12 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import BaseSeries from '../Core/Series/Series.js';
-var PieSeries = BaseSeries.seriesTypes.pie;
-import Chart from '../Core/Chart/Chart.js';
-import H from '../Core/Globals.js';
+import BaseSeries from '../../Core/Series/Series.js';
+var _a = BaseSeries.seriesTypes, LineSeries = _a.line, PieSeries = _a.pie;
+import Chart from '../../Core/Chart/Chart.js';
+import H from '../../Core/Globals.js';
 var noop = H.noop;
-import LineSeries from './Line/LineSeries.js';
-import U from '../Core/Utilities.js';
+import U from '../../Core/Utilities.js';
 var addEvent = U.addEvent, extend = U.extend, fireEvent = U.fireEvent, isArray = U.isArray, merge = U.merge, pick = U.pick;
 /**
  * @private
