+++ js/masters/modules/funnel.src.js
@@ -10,4 +10,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/FunnelSeries.js';
+import '../../Series/Funnel/FunnelSeries.js';
